pyproject.toml
[project]
name = "rl-semantic-trajectories"
version = "0.1.0"
description = "Code for the semantic-aware RL trajectory paper"
authors = [
  { name = "Tobias Kietreiber", email = "tobias.kietreiber@fhstp.ac.at" },
  { name = "Alessio Molinari", email = "alessio.molinari@strg.at" }
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
  "duckdb >=1.0.0,<2.0.0",
  "numpy >=2.0.1,<3.0.0",
  "numba >=0.60.0,<1.0.0",
  "gymnasium >=0.29.1,<1.0.0",
  "igraph >=0.11.6,<1.0.0",
  "tqdm >=4.66.5,<5.0.0",
  "umap-learn >=0.5.6,<1.0.0",
  "pandas >=2.2.2,<3.0.0",
  "plotly >=5.23.0,<6.0.0",
  "cffi >=1.17.0,<2.0.0",
  "stable-baselines3 >=2.3.2,<3.0.0",
  "tensorboard >=2.17.1,<3.0.0",
  "rich >=13.8.0,<14.0.0",
  "pot >=0.9.4,<1.0.0",
  "sb3-contrib >=2.3.0,<3.0.0",
  "python-dotenv >=1.0.1,<2.0.0",
  "transformers (>=4.51.3,<5.0.0)",
  "sentence-transformers (>=4.1.0,<5.0.0)",
  "streamlit>=1.45.0",
  "hdbscan>=0.8.40",
  "tinygrad>=0.10.2",
  "seaborn>=0.13.2",
]

[dependency-groups]
dev = [
    "basedpyright>=1.28.4",
    "debugpy>=1.8.14",
    "ipython>=9.0.2",
    "ruff>=0.11.2",
]
tinygrad = [
    "tinygrad",
]
clustering = [
    # If you're running on GPU, compile this from source. See FAISS' Github page
    "faiss-cpu>=1.11.0",
]
experimental = [
    "ty>=0.0.1a10",
]

[project.optional-dependencies]
cpu = [
  "torch>=2.8.0",
]
cuda = [
  "torch>=2.8.0",
]
rocm = [
  "torch>=2.8.0",
  "pytorch-triton-rocm>=3.4.0"
]

[tool.uv]
conflicts = [
  [
    { extra = "cpu" },
    { extra = "cuda" },
    { extra = "rocm" },
  ],
]

[tool.uv.sources]
torch = [
  { index = "pytorch-cpu", extra = "cpu" },
  { index = "pytorch-rocm", extra = "rocm" },
]
pytorch-triton-rocm = [{ index = "pytorch-rocm" }]
tinygrad = {git = "https://github.com/tinygrad/tinygrad.git"}

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

[[tool.uv.index]]
name = "pytorch-rocm"
url = "https://download.pytorch.org/whl/rocm6.4"
explicit = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
ignore = ["E501"]
extend-select = ["I", "ICN"]
line-length = 120

[tool.basedpyright]
reportMissingTypeStubs = false
typeCheckingMode = "standard"
include = ["bin/", "rl_semantic_trajectories", "evaluation", "experiments"]

[tool.black]
line-length = 120
